requirements.txt
gunicorn
pecan
jsonschema
pika
requests
libvirt-python
jinja2